move TextName into TextSettings since it makes more sense there
