+++ client/anoixo-client/src/query/tests/AttributeQueryCache.test.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { AttributeQueryCacheProvider } from '../AttributeQueryCache';
-import { TextContextProvider, TextName } from '../../TextContext';
+import { TextContextProvider } from '../../TextContext';
+import { TextName } from '../../TextSettings';
 import fetchMock from "jest-fetch-mock"
 import { act } from 'react-dom/test-utils';
